% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunc.R
\name{FDcalc}
\alias{FDcalc}
\title{calculate FD (functional difference) value}
\usage{
FDcalc(df1, df2)
}
\arguments{
\item{df1}{a dataframe with column called FA}

\item{df2}{a dataframe with column called FA}
}
\value{
return the FD value of \code{df1$FA} and \code{df2$FA}
}
\description{
calculate FD (functional difference) value
}

